refactor: Refactored ListToolbar component to include a button with a PlusIcon and moved toggleModal logic to a separate function
