+++ src/components/adminPanel/ListToolBar.jsx
@@ -1,3 +1,7 @@
+import { PlusIcon } from "@heroicons/react/24/outline";
+import { useDispatch } from "react-redux";
+import { toggleModal } from "../../features/userSlice";
+
 const ListToolbar = ({
   title,
   searchTerm,
@@ -5,6 +9,12 @@ const ListToolbar = ({
   placeholder = "Search...",
   actions,
 }) => {
+  const dispatch = useDispatch();
+
+  const handleToogleModal = (payload) => {
+    dispatch(toggleModal(payload));
+  };
+
   return (
     <div className="flex flex-col gap-4 mb-6 md:flex-row md:items-center md:justify-between">
       <h2 className="text-lg font-bold text-navy">{title}</h2>
@@ -22,7 +32,15 @@ const ListToolbar = ({
           </div>
         )}
 
-        <div className="flex gap-2">{actions}</div>
+        <div className="flex gap-2">
+          <button
+            onClick={() => handleToogleModal(true)}
+            className="px-4 py-2 text-white transition duration-300 rounded-md bg-seaGreen hover:bg-navy"
+          >
+            <PlusIcon className="inline-block w-5 h-5 mr-2" />
+            {actions}
+          </button>
+        </div>
       </div>
     </div>
   );
