+++ src/pages/UserManagement.jsx
@@ -1,6 +1,5 @@
 import { useEffect, useState } from "react";
-import { PlusIcon } from "@heroicons/react/24/outline";
-import { useDispatch, useSelector } from "react-redux";
+import { useSelector } from "react-redux";
 import useUserCall from "../hooks/useUserCall";
 import UserRow from "../components/adminPanel/UserRow";
 import UserForm from "../components/adminPanel/UserForm";
@@ -10,16 +9,10 @@ import ListToolbar from "../components/adminPanel/ListToolBar";
 import Pagination from "../components/adminPanel/Pagination";
 
 const UserManagement = () => {
-  const dispatch = useDispatch();
   const { getAllUsers } = useUserCall();
   const { isModalOpen, users } = useSelector((state) => state.users);
   const { pagUsers } = useSelector((state) => state.pagination);
-
   const [searchTerm, setSearchTerm] = useState("");
-
-  const handleToogleModal = (payload) => {
-    dispatch(toggleModal(payload));
-  };
 
   useEffect(() => {
     getAllUsers();
@@ -34,15 +27,7 @@ const UserManagement = () => {
             searchTerm={searchTerm}
             onSearchChange={setSearchTerm}
             placeholder="Search user..."
-            actions={
-              <button
-                onClick={() => handleToogleModal(true)}
-                className="px-4 py-2 text-white transition duration-300 rounded-md bg-seaGreen hover:bg-navy"
-              >
-                <PlusIcon className="inline-block w-5 h-5 mr-2" />
-                Add a new User
-              </button>
-            }
+            actions="Add a new User"
           />
           <div className="space-y-4">
             <div className="hidden text-sm font-medium text-gray-500 md:grid md:grid-cols-12 md:gap-4">
